==> vignettes/04_run_balance_optimisation_PEMmanuscript.Rmd <==
---
title: "03_best_balance"
author: "W.H. MacKenzie"
date: "2023-03-21"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r setup}

devtools::load_all("D:\\GitHub\\PEMprepr")
devtools::load_all("D:\\GitHub\\PEMsamplr")
devtools::load_all("D:\\GitHub\\PEMmodelr")

library(PEMprepr)
library(PEMsamplr)
library(PEMmodelr)
library(dplyr)
library(sf)

```

## Run models

```{r cars}
# set up the standard inputs 
fid <- setup_folders("Deception_AOI")

in_dir <- fid$model_inputs0310[2]

fmat <- read.csv(file.path(in_dir, "fuzzy_matrix.csv"))%>%
  dplyr::select(target, Pred, fVal)

# set up tuning (applies to all models)
best_tune <- fread(file.path(in_dir, "best_tuning.csv"))
mtry <- best_tune$mtry
min_n <- best_tune$min_n

# select reduced variables
reduced_vars <- read.csv(file.path(in_dir,  "reduced_covariate_list.csv")) %>% dplyr::pull()

bgc_pts_subzone <- readRDS(file.path(fid$model_inputs0310[2], "model_input_pts.rds"))

```

## Run Balance Grid search  

This performs a detailed grid search of all the downsampling and smoting options. 

# Alternatively Run the balance optimised version bayesian version (this is currently in development)

```{r}

bal_bgc <- lapply(names(bgc_pts_subzone), function(xx) {
  
  xx <- names(bgc_pts_subzone[1])
  
  alldat = bgc_pts_subzone[[xx]]
  
  outDir = file.path(fid$model_draft[2], xx)
  
  tdat <- alldat %>% mutate(slice = factor(slice))
  tdat <- tdat %>%
    dplyr::select(id,mapunit1, mapunit2, position,
      transect_id, tid, slice, any_of(reduced_vars))
  
  tdat <- tdat[complete.cases(tdat[, 8:length(tdat)]), ]
  tdat <- droplevels(tdat)
  
  balance_optimisation_iteration(
      train_data = tdat,
      fuzz_matrix = fmat,
      ds_iterations = c(10,20,30,40,50,60,70,80,90),
      smote_iterations = c(0.1, 0.2, 0.3),#, 0.4, 0.5, 0.6, 0.7,0.8, 0.9),
      mtry = mtry,
      min_n = min_n,
      use.neighbours = TRUE,
      out_dir = outDir,
      detailed_output = FALSE
  )
  
})

```


# Consolidate all balancing options and find optimum

```{r consolidate acc outputs and graph, echo = FALSE, fig.width=8, fig.height=8}

# combine all balance options into a single data table
acc_total <- foreach::foreach(xx = names(bgc_pts_subzone), .errorhandling = "pass",.combine = rbind) %do% {
  #k = data_list[2]
  # print(k)
  outDir = file.path(fid$model_draft[2], xx)
  allacc <- combine_balance_ouputs(outDir)
  allacc <- allacc %>% dplyr::mutate(bgc = xx)
  
}

# select the best metrics for each Bec variant 

bgcs <- unique(as.factor(acc_total$bgc)) %>% droplevels()

best_results <- foreach(b = levels(bgcs), .combine=rbind) %do% {
  #b <- bgcs[3]
  
  acc_bgc <- acc_total %>% dplyr::filter(bgc %in% b)
  best_metrics <- select_best_acc(acc_bgc) %>% 
    mutate(bgc = b)
  
  best_metrics
  
}

in_dir <- fid$model_inputs0310[2]

write.csv(best_results, file.path(in_dir, "best_balancing.csv"))

```




Alternatively you can run a balance optimize option (see below) 


#  Run the balance optimised version (baysian version)

```{r}
# # # data lines
  train_data <- trDat %>%
    filter(slice != 1)#%>%
    #dplyr::select(-bgc_cat,-transect_id)

  train_data = as.data.table(trDat)
  train_data <-  train_data[grep("ICHmc2_",mapunit1),]
  train_data <- droplevels(train_data)

  fuzz_matrix = fmat
  num_slice = 4
   n_iters = 4
   use.neighbours = TRUE
   acc_mets = c("spat_paf_theta.5")
   mtry = 4 #mtry
   min_n = 7# min_n
   # end data lines

#metrics_to_run <- c("spat_paf_theta.5")

opt_res <- optimise_balance(train_data, 
                 fuzz_matrix, 
                 num_slice = 4, 
                 n_iters = 4, 
                 mtry = 14,
                 min_n = 7, 
                 use.neighbours = TRUE,
                 acc_mets = acc_mets)


# get best balance: 
getBestPars(opt_res, N = 5)

```
